requirements.txt
plyfile
h5py
lightly
torch_optimizer
open3d